worker/requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.2
redis==5.0.1
rq==1.15.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
lxml==4.9.3
requests==2.31.0
